requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Flask-WTF==0.8.3
Jinja2==2.7
MarkupSafe==0.18
SQLAlchemy==0.8.1
WTForms==1.0.4
Werkzeug==0.8.3
argparse==1.2.1
distribute==0.6.34
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
mplayer.py==0.7.0
wsgiref==0.1.2
